Move InstallAppRpc
