--- os/server/computers/real/src/jvmMain/kotlin/com/wasmo/computers/ComputerBindings.kt
@@ -1,11 +1,0 @@
-package com.wasmo.computers
-
-import dev.zacsweers.metro.BindingContainer
-import dev.zacsweers.metro.Binds
-
-@BindingContainer
-abstract class ComputerBindings {
-
-  @Binds
-  abstract fun bindComputerStore(real: RealComputerStore): ComputerStore
-}
