+++ os/server/computers/real/src/jvmMain/kotlin/com/wasmo/computers/ComputersActions.kt
@@ -3,6 +3,7 @@ package com.wasmo.computers
 import com.wasmo.framework.UserAgent
 
 interface ComputersActions {
+  val installAppRpc: InstallAppRpc
   val createComputerSpecRpc: CreateComputerSpecRpc
 
   interface Factory {
