+++ os/server/computers/real/src/jvmMain/kotlin/com/wasmo/computers/InstallAppRpc.kt
@@ -1,10 +1,9 @@
-package com.wasmo.installedapps
+package com.wasmo.computers
 
 import com.wasmo.accounts.CallScope
 import com.wasmo.accounts.Client
 import com.wasmo.api.InstallAppRequest
 import com.wasmo.api.InstallAppResponse
-import com.wasmo.computers.ComputerStore
 import com.wasmo.framework.NotFoundUserException
 import com.wasmo.framework.Response
 import com.wasmo.identifiers.AppSlugRegex
