--- os/server/installedapps/real/src/jvmMain/kotlin/com/wasmo/installedapps/ComputerActions.kt
@@ -1,12 +1,0 @@
-package com.wasmo.installedapps
-
-import com.wasmo.framework.UserAgent
-
-interface ComputerActions {
-  val installAppRpc: InstallAppRpc
-
-  interface Factory {
-    fun create(userAgent: UserAgent): ComputerActions
-  }
-}
-
