+++ os/server/testing/src/jvmMain/kotlin/com/wasmo/testing/service/ServiceTesterGraph.kt
@@ -2,8 +2,8 @@ package com.wasmo.testing.service
 
 import com.wasmo.accounts.AccountsBindings
 import com.wasmo.accounts.ClientAuthenticator
-import com.wasmo.computers.ComputerBindings
 import com.wasmo.computers.ComputerServiceGraph
+import com.wasmo.computers.ComputersBindings
 import com.wasmo.identifiers.Deployment
 import com.wasmo.identifiers.OsScope
 import com.wasmo.installedapps.InstalledAppBindings
@@ -39,7 +39,7 @@ import wasmo.time.FakeClock
   bindingContainers = [
     AbsurdBindings::class,
     AccountsBindings::class,
-    ComputerBindings::class,
+    ComputersBindings::class,
     InstalledAppBindings::class,
     PasskeysBindings::class,
     PermitsBindings::class,
